refactor: refactor application middleware
